Renamed test.go to goclient_test.go. Started converting code into unit test suite. Ledger's signature validation is working using go's crypto.
