+++ src/goclient/Gopkg.lock
@@ -243,6 +243,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "b210e3d634342ffc427364fb952ca7c693d439e74899ce4697e3d81152e17b9e"
+  inputs-digest = "2781169ec6aa778ef69fbe2be795018f2a12734eb97b16c6ae53e15e9c9e6c80"
   solver-name = "gps-cdcl"
   solver-version = 1
